pyproject.toml
[project]
name = "a2a-experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[http-server]>=0.3.3",
    "asyncio-mqtt>=0.16.2",
    "docker>=7.1.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "psutil>=7.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "redis>=6.4.0",
    "uvicorn>=0.35.0",
]
